backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
django-cors-headers==4.3.1
celery==5.3.6
redis==5.0.3
psycopg2-binary==2.9.9
pgvector==0.2.5
gitpython==3.1.43
google-generativeai==0.7.2
python-dotenv==1.0.1
ruff==0.4.4
